+++ packages/sdk/src/Incentives/ERC20VariableCriteriaIncentiveV2.ts
@@ -457,7 +457,7 @@ export function prepareERC20VariableCriteriaIncentiveV2Payload({
             components: [
               { type: 'uint8', name: 'criteriaType' },
               { type: 'bytes32', name: 'signature' },
-              { type: 'uint8', name: 'fieldIndex' },
+              { type: 'uint32', name: 'fieldIndex' },
               { type: 'address', name: 'targetContract' },
               { type: 'uint8', name: 'valueType' },
             ],
